refactor(edge): isolate upstream dispatch retry and hedging logic
